Extract PrimitivesLogProcessor module
