+++ lib/log-parsers.js
@@ -2,15 +2,6 @@
 
 const jmespath = require('jmespath')
 
-function primitives (input) {
-  // Short-circuit for spec allowed primitive values.
-  if ([null, true, false].includes(input) || Number.isFinite(input)) {
-    return { output: `${input}\n`, done: true }
-  } else {
-    return { output: input }
-  }
-}
-
 function search (input, { search }) {
   if (search && !jmespath.search(input, search)) {
     return { output: undefined, done: true }
@@ -35,7 +26,6 @@ function ignore (input, { ignore }) {
 }
 
 const builtInLogParsers = {
-  primitives,
   search,
   ignore
 }
