+++ lib/log-processors.js
@@ -3,8 +3,9 @@ const { prettifyLevel, prettifyMessage, prettifyMetadata, prettifyTime } = requi
 const { builtInLineBuilders } = require('./line-builders')
 const LogProcessor = require('./LogProcessor')
 const JsonLogProcessor = require('./processors/JsonLogProcessor')
+const PrimitivesLogProcessor = require('./processors/PrimitivesLogProcessor')
 
-const { primitives, search, ignore } = builtInLogParsers
+const { search, ignore } = builtInLogParsers
 const {
   appendLevel,
   appendMessage,
@@ -27,12 +28,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class PrimitivesLogProcessor extends LogProcessor {
-  parse (input, context) {
-    return primitives(input, context)
   }
 }
